@@ -1,24 +1,20 @@
-const {DymoServices, createImageWithText} = require('node-dymo-printer');
+import {createImageWithText, DymoServices} from 'node-dymo-printer';
 
 /**
- * Try to find the DYMO LabelWriter and print "Hello World!" on a 89mm x 36mm label.
+ * Try to find the DYMO LabelWriter and print "Hello World".
  */
 
-(async function () {
+// Create a landscape image with the dimensions of the label and with the text "Hello World".
+const {imageWidth, imageHeight} = DymoServices.DYMO_LABELS['54mm x 25mm'];
+const image = await createImageWithText(imageWidth, imageHeight, 50, 128, 'Hello World!');
 
-    // Create landscape image with the dimensions of the label and with the text "Hello World!".
-    const {imageWidth, imageHeight} = DymoServices.DYMO_LABELS['89mm x 36mm'];
-    const image = await createImageWithText(imageWidth, imageHeight, 50, 128, 'Hello World!');
+// For debugging purposes, write the image to disk.
+image.write('./image1.png');
 
-    // For debugging purposes, write the image to disk.
-    image.write(__dirname + '/image1.png');
-
-    // Print it, just one label.
-    try {
-        await new DymoServices().print(image, 1);
-        console.log('Successfully printed');
-    } catch (e) {
-        console.error('Error while trying to print the image: ', e);
-    }
-
-})();
+// Print it, just one label.
+try {
+    await new DymoServices().print(image, 1);
+    console.log('Successfully printed');
+} catch (e) {
+    console.error(e);
+}
